+++ src/GA_core/gene.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+
+class Gene(ABC):
+
+    @abstractmethod
+    def mutate_gene(self):
+        pass
